django_project/requirements.txt
django
django-rest-framework
djoser
djangorestframework-simplejwt
django-cors-headers
django-environ
pymysql
python-dotenv
lxml
bleach
python-dotenv